% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_fgeo.R
\name{boot_fgeo}
\alias{boot_fgeo}
\title{Bootstrap values to get site-level estimates and 95\% confidence intervals}
\usage{
boot_fgeo(
  value,
  weight,
  subplot,
  variable,
  group,
  group2,
  year = NULL,
  dT = NULL,
  kohyama = NULL,
  nrep = 1000,
  para = FALSE
)
}
\arguments{
\item{value}{Data to be bootstraped, as a numerical vector.}

\item{weight}{Weight given to observations, as a numerical vector of same
length as \code{value}.}

\item{subplot}{Subplots observations belong to.}

\item{variable}{Variables observations belong to.}

\item{group}{Grouping factor (for example size).}

\item{group2}{Additional grouping factor (for example site).}

\item{year}{Year observations belong to. Default is \code{NULL}.}

\item{dT}{Time interval (years) between the current census (as in \code{year}) and
the next one.}

\item{kohyama}{Name of variables to apply Kohyama correction (2019), should
be a vector of length 3 with name of stock variable, influx variable, and
outflux variable (in that order). Default is \code{NULL} (i.e. correction not
applied).}

\item{nrep}{Default is 1000.}

\item{para}{Default is \code{FALSE}.}
}
\value{
A data frame with relevant variables.
}
\description{
Bootstrap values to get site-level estimates and 95\% confidence intervals
}
