% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chojnackyParams.R
\name{chojnackyParams}
\alias{chojnackyParams}
\title{Retrieve Chojnacky parameters based on taxonomy and wood density}
\usage{
chojnackyParams(family, genus, wsg)
}
\arguments{
\item{family}{Character vector containing name of family corresponding to observations}

\item{genus}{Character vector containing genus}

\item{wsg}{Numerical vector containing wood density}
}
\value{
A matrix with 2 columns: the intercept and slope in the Chojnacky equations (Chojnacky et al, 2014)
}
\description{
Retrieve Chojnacky parameters based on taxonomy and wood density
}
