% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_spptab.R
\name{clean_spptab}
\alias{clean_spptab}
\title{Open and clean species table.}
\usage{
clean_spptab(path_spp = getwd())
}
\arguments{
\item{path_spp}{The directory were the species table(s) are. Species tables must
be .Rdata files, and contain 'spp' in the name. Columns must contain: xxx}
}
\value{
A data table with the following columns: "site" (site name, as spelled in the .Rdata file), "sp" (species code), "name" (Latin name), "taxo_group" (palm, vine, tree fern, gymnosperm, or NA for other groups)
}
\description{
Open and clean species table.
}
