% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate_data.R
\name{consolidate_data}
\alias{consolidate_data}
\title{ForestGEO-like data consolidation}
\usage{
consolidate_data(
  df,
  taper_correction = TRUE,
  stem_matching = TRUE,
  add_missing = TRUE,
  correct_diam = TRUE,
  step_corr = TRUE,
  dcor_min = 100,
  acc_decr = -5,
  acc_incr = 50,
  relat_change = FALSE,
  species_path = "data/species/"
)
}
\arguments{
\item{df}{A data.table obtained with the prepare_data() function.}

\item{taper_correction}{\code{TRUE} (default) or \code{FALSE}, apply Cushman et al (2014) taper
correction.}

\item{stem_matching}{\code{TRUE} (default) or \code{FALSE}, apply the stem matching script.}

\item{add_missing}{\code{TRUE} (default) or \code{FALSE}, interpolate missing stem measurements}

\item{correct_diam}{\code{TRUE} (default) or \code{FALSE}, should dbh errors (detected when dbh changes are
outside the acceptable range) be corrected?}

\item{step_corr}{\code{TRUE} (default) or \code{FALSE}, should step dbh changes be corrected?}

\item{acc_decr}{numeric: acceptable decrease between two censuses, as a proportion
if relat_change is TRUE or in mm if relat_change is false}

\item{acc_incr}{numeric: acceptable annual increase, as a proportion (per year)
if relat_change is TRUEor in mm/year if relat_change is false}

\item{relat_change}{logical: should the decrease and increse values be relative
(proportion of total dbh) or absolute values (in mm)?}
}
\value{
A data.table (data.frame) with all relevant variables.
}
\description{
ForestGEO-like data consolidation
}
