% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_subplots.R
\name{create_subplots}
\alias{create_subplots}
\title{Create subplots of fixed size based on tree coordinates}
\usage{
create_subplots(x, y, size = 20, Nmin = 100)
}
\arguments{
\item{x}{x coordinate of each observation}

\item{y}{y coordinate of each observation}

\item{size}{Length of the subplot side (in m; subplots are squares).}

\item{Nmin}{Minimum acceptable abundance (n/ha) under which subplots are discarted. Default is 100.}
}
\value{
A vector with new subplot
}
\description{
Create subplots of fixed size based on tree coordinates
}
