% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_census_year.R
\name{get_census_year}
\alias{get_census_year}
\title{Extract census year from a set of exact measure dates}
\usage{
get_census_year(date)
}
\arguments{
\item{date}{vector of exact measurement dates, containing the year,
month and day (not always in this order) and sometimes
the hour of all measurements during one census}
}
\description{
Extract census year from a set of exact measure dates
}
