% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_stems.R
\name{match_stems}
\alias{match_stems}
\title{Matching unidentifed stems of one tree in ForestGEO data}
\usage{
match_stems(dbh, year, acc_decr = -5, acc_incr = 35, relat_change = FALSE)
}
\arguments{
\item{dbh}{numeric vector containing dbh values}

\item{year}{numeric vector (same size as dbh) containing each measure census year}

\item{acc_decr}{numeric: acceptable decrease between two censuses, as a proportion if relat_change is TRUE
or in mm if relat_change is false}

\item{acc_incr}{numeric: acceptable annual increase, as a proportion (per year) if relat_change is TRUE
or in mm/year if relat_change is false}

\item{relat_change}{logical: should the decrease and
increse values be relative (proportion of total dbh) or absolute values (in mm)?}
}
\value{
numeric vector of the same length as dbh and year, with the stem ID of each measurement
}
\description{
Matching unidentifed stems of one tree in ForestGEO data
}
