% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_FGEOdata.R
\name{open_FGEOdata}
\alias{open_FGEOdata}
\title{Open ForestGEO .Rdata census data.}
\usage{
open_FGEOdata(file)
}
\arguments{
\item{file}{character string - name of the Rdata file to open (with its path). The name should be written
as path/site.stem#census.rdata, or path/site.full#census.rdata
(eg: data/bci.stem2.rdata is the 2nd census at bci, containing all stems)}
}
\value{
A data frame (or a list of data frames) with the census data.
}
\description{
Open ForestGEO .Rdata census data.
}
