% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppa.R
\name{ppa}
\alias{ppa}
\title{Estimate canopy position of trees based on perfect plasticity approximation}
\usage{
ppa(ca, area)
}
\arguments{
\item{ca}{A numerical vector containing the crown area of indivudual trees
(in ha)}

\item{area}{Area of the plot}
}
\value{
A vector with logical values, \code{TRUE} if the tree is in the canopy
(according ot the PPA) and \code{FALSE} otherwise.
}
\description{
Estimate canopy position of trees based on perfect plasticity approximation
}
