% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Import and homogenize ForestGEO-like data.}
\usage{
prepare_data(
  path,
  site,
  stem = TRUE,
  dbh_units = "mm",
  exclude_interval = NULL,
  keep_columns = c("treeid", "stemid", "tag", "stemtag", "sp", "quadrat", "gx", "gy",
    "dbh", "hom", "exactdate", "codes", "dfstatus", "agb", "brokenladder", "fire.code")
)
}
\arguments{
\item{path}{String giving a path to a parent directory containing species
and census datasets.}

\item{site}{String giving the name of your site -- one of one of
\code{site_info$site} (e.g., 'barro colorado island'.}

\item{stem}{\code{TRUE}(default) or \code{FALSE} reflect that your censuses are
ForestGEO \code{stem} or \code{tree} (aka \code{full}) tables, respectively.}

\item{dbh_units}{set the unit ("mm" or "cm") of DBH values, by default
dbh_units=="mm".}

\item{exclude_interval}{\code{NULL} by default. If needed a vector (e.g. c(1,2))
indicating which census interval(s) must be discarded from computation due,
for instance, to a change in measurement protocol.}

\item{keep_column}{character vector containing the name of the columns to keep
in the data. Default is: c("treeid", "stemid", "tag", "stemtag",
"sp", "quadrat", "gx", "gy", "dbh",
"hom", "exactdate", "codes", "dfstatus",
"agb", "brokenladder", "fire.code")}
}
\value{
A data.table (data.frame) with all relevant variables.
}
\description{
Main routine to format, detect major obvious errors, and gap-fill those
errors in ForestGEO-like data.
}
