% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stand_dynamics.R
\name{stand_dynamics}
\alias{stand_dynamics}
\title{Calculation of stand-level dynamics}
\usage{
stand_dynamics(
  id,
  var,
  dvar = NULL,
  year,
  group = NULL,
  plot = NULL,
  site = NULL
)
}
\arguments{
\item{id}{A vector of individual trees (or stems) IDs}

\item{var}{A numerical vector of the same size as id, containing the
measurements of the target variable}

\item{dvar}{Optional, a vector with variable change between 2 censuses.}

\item{year}{A numerical vector (same size as id), containing census years}

\item{group}{Optional; a vector (same size as id), containing a grouping factor.}

\item{site}{Optional; a vector (same size as id), containing the measurements' site}
}
\value{
A data.table (data.frame) with stocks, annual gain and loss of var
for every census year and group level.
}
\description{
Calculation of stand-level dynamics
}
