% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substitute_change.R
\name{substitute_change}
\alias{substitute_change}
\title{Substitute DBH changes (cm/yr) outside of a pre-set acceptable range of DBH
changes by their expected values based on other observed values inside the
pre-set interval. Can also substitute AGB changes by intergrating the AGB
allometry over the distribution of expected DBH changes. This function allows
to have unbiased stand-level estimates of DBH and AGB growth while removing
the most obvious measurement errors.}
\usage{
substitute_change(
  varD,
  hom_change,
  cut = c(-0.5, 5),
  lambda = 0.5,
  value = "D",
  D = NULL,
  WD = NULL,
  E = NULL,
  a = NULL,
  b = NULL
)
}
\arguments{
\item{varD}{Vector of changes in DBH.}

\item{hom_change}{Vector of changes in height of measurement.}

\item{cut}{Acceptable range of DBH changes. Default is c(-0.5, 5)}

\item{lambda}{Parameter of the modulus function used to normalize DBH changes
(see Condit et al., 2017). Default is 0.5}

\item{value}{What change should be computed, either "D" (DBH) or "AGB".
Default is "D".}

\item{D}{Must provide a value when value = "AGB". Default is NULL.}

\item{WD}{Vector of wood densities. Must provide a value when value = "AGB"
(can be "NA" for temperate sites). Default is NULL.}

\item{E}{Vector of environmental parameter used in Chave et al pantropical
allometric equation. Must provide a value when value = "AGB" (NA for
temperate sites). Default is NULL.}

\item{a}{Vector of first parameter used in Chojnacky allometric equations.
Must provide a value when value = "AGB" (NA for tropical sites). Default is
NULL.}

\item{b}{Vector of second parameter used in Chojnacky allometric equations.
Must provide a value when value = "AGB" (NA for tropical sites). Default is
NULL.}
}
\value{
A vector of corrected changes in DBH (if value = "D) or AGB (if value
= "AGB).
}
\description{
Substitute DBH changes (cm/yr) outside of a pre-set acceptable range of DBH
changes by their expected values based on other observed values inside the
pre-set interval. Can also substitute AGB changes by intergrating the AGB
allometry over the distribution of expected DBH changes. This function allows
to have unbiased stand-level estimates of DBH and AGB growth while removing
the most obvious measurement errors.
}
