% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_corrections.R
\name{visualize_corrections}
\alias{visualize_corrections}
\title{Visualize corrections made by the consolidate_data() function}
\usage{
visualize_corrections(df, dbhmin = 500, plotsPerPage = 25)
}
\arguments{
\item{df}{data.table: The output of the function consolidate_data()}

\item{dbhmin}{numeric: the minimum dbh of trees to plot}

\item{plotsPerPage}{numeric: number of plots (stemids) per page}
}
\value{
A list of ggplots
}
\description{
Visualize corrections made by the consolidate_data() function
}
