% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_census_year.R
\name{get_census_year}
\alias{get_census_year}
\title{Extract census year from date column}
\usage{
get_census_year(date)
}
\arguments{
\item{date}{A vector containing date information (accepts several formats: DD-MM-YYYY, MM-DD-YYYY, YYYY-MM-DD, YY-MM-DD, etc)}
}
\value{
A vector containing correponding census year.
}
\description{
Extract census year from date column
}
